% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_summary.funct.R
\name{aovInteractSummaryTable}
\alias{aovInteractSummaryTable}
\title{AoV summary table with interaction}
\usage{
aovInteractSummaryTable(
  aov_data,
  group_var_1,
  group_var_2,
  interaction_funct = "*",
  output_file = "stats_interact_summary.xlsx"
)
}
\arguments{
\item{aov_data}{The data frame containing the data to be analysed.}

\item{group_var_1}{The name of the column containing the first group names (in quotes).}

\item{group_var_2}{The name of the column containing the second group names (in quotes).}

\item{interaction_funct}{The interaction function to use. Defaults to "*".}

\item{output_file}{The name of the output file (in quotes). Defaults to "stats_summary.xlsx".}
}
\value{
Returns a summary table and saves it as an Excel file in the active directory.
}
\description{
This allows you to generate a table of means and key statistical
 values from an analysis of variance (AoV) test for a whole dataframe,
 with interactions between two factors. 
 It uses the agricolae, dplyr, gtools, and openxlsx packages.\cr
 Please convert any non-numeric columns to factors before using this function.
 Ideally, factor columns should be the first few columns in the data frame.\cr
}
\examples{
aovSummaryTable(aov_data, "Group", "stats_summary.xlsx")
}
