% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_summary.funct.R
\name{aovInteractSummaryTable}
\alias{aovInteractSummaryTable}
\title{AoV summary table with interaction}
\usage{
aovInteractSummaryTable(
  aov_data,
  group_var_1,
  interaction_funct = "*",
  group_var_2,
  output_file = NULL
)
}
\arguments{
\item{aov_data}{The data frame containing the data to be analysed.}

\item{group_var_1}{The name of the column containing the first group names (in quotes).}

\item{interaction_funct}{The interaction function to use. Defaults to "*".}

\item{group_var_2}{The name of the column containing the second group names (in quotes).}

\item{output_file}{If a file name is supplied (in quotes), it will output an Excel file with the summary table. The filename should have an .xlsx extension.}
}
\value{
Returns a summary table and (optionally) saves it as an Excel file in the active directory.
}
\description{
This allows you to generate a table of means and key statistical
 values from an analysis of variance (AoV) test for a whole dataframe,
 with interactions between two factors. 
 It uses the agricolae, dplyr, gtools, and openxlsx packages.\cr
 Please convert any non-numeric columns to factors before using this function.
 Ideally, factor columns should be the first few columns in the data frame.\cr
}
\examples{
aovSummaryTable(aov_data, "Group1","*","Group2", "stats_summary.xlsx")
}
