% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_summary.funct.R
\name{aovSummaryTable}
\alias{aovSummaryTable}
\title{AoV summary table}
\usage{
aovSummaryTable(aov_data, group_var, output_file = "stats_summary.xlsx")
}
\arguments{
\item{aov_data}{The data frame containing the data to be analysed.}

\item{group_var}{The name of the column containing the group names (in quotes).}

\item{output_file}{The name of the output file (in quotes). Defaults to "stats_summary.xlsx".}
}
\value{
Returns a summary table and saves it as an Excel file in the active directory.
}
\description{
This allows you to generate a table of means and key statistical
 values from an analysis of variance (AoV) test. It uses the agricolae, dplyr,
 gtools, and openxlsx packages.
 Please convert any non-numeric columns to factors before using this function.
 Ideally, factor columns should be the first few columns in the data frame. 
 It does not cope with interactions at the moment.
}
\examples{
aovSummaryTable(aov_data, "Group", "stats_summary.xlsx")
}
