% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cielab.funct.R
\name{deltaE76}
\alias{deltaE76}
\title{Calculate CIE76 colour difference (Delta E)}
\usage{
deltaE76(L1, a1, b1, L2, a2, b2)
}
\arguments{
\item{L1, a1, b1}{CIELab coordinates of first colour}

\item{L2, a2, b2}{CIELab coordinates of second colour}
}
\value{
Numeric Delta E value(s)
}
\description{
Computes the Euclidean distance between two colours in CIELab space,
also known as Delta E (CIE 1976).
}
\details{
Interpretation of Delta E values:
\itemize{
  \item 0-1: Not perceptible by human eye
  \item 1-2: Perceptible through close observation
  \item 2-10: Perceptible at a glance
  \item 11-49: Colours are more similar than opposite
  \item 100: Colours are exact opposites
}
}
\examples{
deltaE76(50, 10, 20, 55, 15, 25)
}
